extract getLeaveWorkspaceConfirmationPrompt into WorkspacesSettingsUtils
