+++ src/libs/WorkspacesSettingsUtils.ts
@@ -6,6 +6,7 @@ import CONST from '@src/CONST';
 import type {TranslationPaths} from '@src/languages/types';
 import type {Policy, ReportAttributesDerivedValue} from '@src/types/onyx';
 import type {CompanyAddress, Unit} from '@src/types/onyx/Policy';
+import {getConnectionExporters, isPolicyAdmin, isPolicyApprover, isPolicyAuditor} from './PolicyUtils';
 
 type BrickRoad = ValueOf<typeof CONST.BRICK_ROAD_INDICATOR_STATUS> | undefined;
 
@@ -135,5 +136,43 @@ function getOwnershipChecksDisplayText(
     return {title, text, buttonText};
 }
 
-export {getChatTabBrickRoadReportID, getBrickRoadForPolicy, getChatTabBrickRoad, getUnitTranslationKey, getOwnershipChecksDisplayText, getWorkspaceAddressStreetLines};
+function getLeaveWorkspaceConfirmationPrompt(policy: OnyxEntry<Policy>, userEmail: string, ownerDisplayName: string, translate: LocaleContextProps['translate']): string {
+    const exporters = getConnectionExporters(policy);
+
+    if (policy?.achAccount?.reimburser === userEmail) {
+        return translate('common.leaveWorkspaceReimburser');
+    }
+
+    if (policy?.technicalContact === userEmail) {
+        return translate('common.leaveWorkspaceConfirmationTechContact', ownerDisplayName);
+    }
+
+    if (exporters.some((exporter) => exporter === userEmail)) {
+        return translate('common.leaveWorkspaceConfirmationExporter', ownerDisplayName);
+    }
+
+    if (isPolicyApprover(policy, userEmail)) {
+        return translate('common.leaveWorkspaceConfirmationApprover', ownerDisplayName);
+    }
+
+    if (isPolicyAdmin(policy)) {
+        return translate('common.leaveWorkspaceConfirmationAdmin');
+    }
+
+    if (isPolicyAuditor(policy)) {
+        return translate('common.leaveWorkspaceConfirmationAuditor');
+    }
+
+    return translate('common.leaveWorkspaceConfirmation');
+}
+
+export {
+    getChatTabBrickRoadReportID,
+    getBrickRoadForPolicy,
+    getChatTabBrickRoad,
+    getUnitTranslationKey,
+    getOwnershipChecksDisplayText,
+    getWorkspaceAddressStreetLines,
+    getLeaveWorkspaceConfirmationPrompt,
+};
 export type {BrickRoad};
